Configurando package.json para o deploy na heroku
