+++ package.json
@@ -4,7 +4,8 @@
   "description": "## Desenvolvedores  - Estéfane Luz - Lucas Goulart",
   "main": "index.js",
   "scripts": {
-    "dev": "nodemon ./src/index.js"
+    "dev": "nodemon ./src/index.js",
+    "start": "node src/index.js"
   },
   "repository": {
     "type": "git",
